# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    API for messaging app WASAText
  version: 0.0.1
security:
  - bearerAuth: [ ]
...
paths:
...
  /chats:
    get:
      tags:
        - chats
      operationId: getMyConversations
      summary: Get user chats
      description: "Get all the user's conversations"
      responses:
        "200":
          description: Chats found
          content:
            application/json:
              schema:
                description: chats
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
                minItems: 1
                maxItems: 999
        "204":
          description: "No Chats"
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - chats
      summary: Create new Chat
      description: |
        Send a new message in a conversation. If it's the first one,
        a new conversation will be created
      operationId: createChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirstMessageRequest"
        required: true
      responses:
        '201':
          description: New chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /chats/{chatId}:
    parameters:
      - $ref: "#/components/parameters/ChatID"
      - name: chatName
        in: query
        description: New group name
        schema:
          type: string
          description: Group name
          example: "Study group"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 3
          maxLength: 24
    get:
      tags:
        - chats
      summary: Get messages in a chat
      description: Returns an array of messages sent in a chat
      operationId: getConversation
      responses:
        "200":
          description: Messages found
          content:
            application/json:
              schema:
                description: chat messages
                type: array
                items:
                  $ref: "#/components/schemas/Message"
                minItems: 1
                maxItems: 99999
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - chats
      summary: Updates the group name
      description: "Updates the group name with the one given in the request"
      operationId: setGroupName
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                chatName:
                  type: string
                  description: Name of the chat to create if new.
                  example: "Project X"

      responses:
        "200":
          description: Chat name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - chats
      summary: Delete chat
      description: "Delete a chat"
      operationId: deleteChat
      responses:
        "204":
          description: Chat successfully deleted
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/chats/{chatId}/messages:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/ChatID"
    post:
      tags:
        - messages
      summary: Send a message
      description: |
        Send a new message in a conversation. If it's the first one,
        a new conversation will be created
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
        required: true
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /users/{userId}/chats/{chatId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/ChatID"
      - $ref: "#/components/parameters/MessageID"
    post:
      tags:
        - messages
      summary: Forward a message
      description: "Get a message to forward it"
      operationId: forwardMessage
      responses:
        "200":
          description: Message found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        "404":
          $ref: "#/components/responses/Response"
    delete:
      tags:
        - messages
      summary: Delete message
      description: "Delete a message"
      operationId: deleteMessage
      responses:
        "200":
          $ref: "#/components/responses/Response"
        "404":
          $ref: "#/components/responses/Response"
  /users/{userId}/chats/{chatId}/messages/{messageId}/comments:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/ChatID"
      - $ref: "#/components/parameters/MessageID"
    get:
      tags:
        - comments
      operationId: getComments
      summary: Get message comments
      description: "Get all the message's comments"
      responses:
        "200":
          description: Comments found
          content:
            application/json:
              schema:
                description: message comments
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
                minItems: 1
                maxItems: 9999
        "204":
          description: "No Comments"
    put:
      tags:
        - comments
      operationId: commentMessage
      summary: Comment a message
      description: "Add a comment to a message. If user's comment is already present, nothing will change"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        "200":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        "204":
          description: Comment already present
...
components:
  schemas:
...
    Chat:
      title: Chat
      description: Generic chat
      type: object
      properties:
        id:
          description: Unique identifier for the chat
          type: integer
          example: 223
          readOnly: True
        isGroup:
          description: identifies chat as a group or private chat
          type: boolean
          example: true
        chatName:
          description: Chat title, user name or group name
          type: string
          example: "Study group"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 3
          maxLength: 24
        lastMsgText:
          description: text of the last message in the chat
          type: string
          example: "The meeting is canceled"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 1
          maxLength: 99
        lastMsgTime:
          description: Date and time when the message is sent
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          example: "2024-10-11T12:14:23Z"
          readOnly: true
          minLength: 20
          maxLength: 20
        lastMsgType:
          description: type of the message, either text or image
          type: string
          enum: ["text", "image"]
          example: "text"
    Message:
      title: Message
      description: Single message
      type: object
      properties:
        id:
          description: Unique identifier for the message
          type: integer
          example: 223
          readOnly: True
        type:
          description: type of the message, either text or image
          type: string
          enum: [ "text", "image" ]
          example: "text"
        text:
          description: message text
          type: string
          example: "hello world"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 1
          maxLength: 250
        chatId:
          description: Identifier for the chat
          type: integer
          example: 223
        senderId:
          description: Identifier for the sender
          type: integer
          example: 223
        isForward:
          type: boolean
          description: Whether this message is forwarded from another chat.
          example: false
        replyTo:
          description: Identifier for the message this one is replying to
          type: integer
          example: 223
        createdAt:
          description: Date and time when the message is sent
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          example: "2024-10-11T12:14:23Z"
          readOnly: true
          minLength: 20
          maxLength: 20
    Comment:
      title: Comment
      description: Single comment
      type: object
      properties:
        userId:
          description: Unique identifier for the comment's user
          type: integer
          example: 223
          readOnly: True
        timestamp:
          description: Date and time when the comment is put
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          example: "2024-10-11T12:14:23Z"
          readOnly: true
          minLength: 20
          maxLength: 20
...
    FirstMessageRequest:
      type: object
      properties:
        text:
          type: string
          description: The text content of the message.
          example: "Hello, team!"
        type:
          type: string
          description: The type of message (e.g., text, image, video).
          example: "text"
        mediaUrl:
          type: string
          format: uri
          description: URL to any attached media.
          example: "https://cdn.example.com/media/abc123.jpg"
        chatName:
          type: string
          description: Name of the chat to create if new.
          example: "Project X"
        receivers:
          type: array
          description: List of user IDs to receive the message.
          items:
            type: integer
          example: [ 123, 456, 789 ]
        isForward:
          type: boolean
          description: Whether this message is forwarded from another chat.
          example: false
      required:
        - text
        - type
        - receivers
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "The request could not be processed"
        code:
          type: integer
          example: 400
  parameters:
    UserID:
      schema:
        type: number
        format: integer
        example: 134
      name: userId
      in: path
      required: true
      description: User identifier
    ChatID:
      schema:
        type: number
        format: integer
        example: 134
      name: chatId
      in: path
      required: true
      description: Chat identifier
    MessageID:
      schema:
        type: number
        format: integer
        example: 134
      name: messageId
      in: path
      required: true
      description: Message identifier
...
  responses:
    Response:
      description: Operation response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                description: Response code
                pattern: "^[0-9]{3}$"
                example: "200"
                minLength: 3
                maxLength: 3
              message:
                type: string
                description: Response message
                pattern: "^[a-zA-Z ]+$"
                minLength: 2
                maxLength: 99
                example: "Successful operation"
    UnauthorizedError:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequestError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer